move tui package to pkg dir
